Many updates to functions
Changed the abs data format to a matrix to match the eems
Created generic methods for subset and unique to streamline some of the utils
Nope this didn’t work well, backing out of this… but the eemlist and abslist are so similar that most of the functions work if it you give it either one. I did get unique to work though, so I think that’s good.
Changed the metaname to meta_name and left sample alone
Updated all the functions so they should run with the revised changes.