QAQC changes to get_sample_info and create_validation_models
get_sample_info now returns a dataframe when asked for an abslist of data. Data.frame
has wavelength column and a bunch of absorbance columns named after the sample name.

This is used in create_validation_model to take a directory of absorbance files,
read them in, and then calculate mean and standard deviation of absorbance for all the
samples.

Also renamed the long term standards directory to long-term-standards (removed spaces)

TODO - where to save the user created absorbance model?