still working on add_metadata
It's difficult to work around the eemlist having renamed instrument blanks for each sample in addition to the samples and sample blanks. Pattern matching is fragile (some samples are actually blanks) and there are double the eems as samples in the metadata so defining is_blank in metadata doesn't work either.