changing tea to check_std or just std to be consistent
when I wrote I switch a lot between std, check std and tea std. The sample type is check, so renaming all the examples and functions to match this naming scheme. Makes sense because in theory this could be used for any check standard, it doesn't have to be tea.